requirements.txt
# MCP (Model Context Protocol) dependencies
mcp==1.12.2

# Database
psycopg2-binary==2.9.10
openai>=1.3.0
# Core dependencies for MCP server
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
typing_extensions==4.14.1

# HTTP and async support
httpx==0.28.1
httpx-sse==0.4.1
anyio==4.9.0
sniffio==1.3.1

# JSON schema support
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
referencing==0.36.2
rpds-py==0.26.0

# Additional utilities
click==8.2.1
certifi==2025.7.14
idna==3.10
charset-normalizer==3.4.2
urllib3==2.5.0
